Add kernel creation check to optimize tab switching greeting
The early return condition for voice check was preventing the
greeting from playing when switching tabs.

Removing the condition made the brain reload everytime we switch
to the chatbot. Which made the process slower.

Adding this check here, avoids reloading the brain everytime we
switch tabs, while also allowing greeting to be played
everytime.